requirements.txt
Mezzanine==3.1.9
celery==3.1.16
mysql-python==1.2.5
django-post_office==1.1.1
django-stripe-payments==2.0.0
